fix(audio): normalize greyScale output to float [0,1] range
The greyScale() function was producing 16-bit values (0-65535) while
the audio pipeline expected normalized float values [0,1]. This caused
massive saturation when summation_response_exponent > 1.

Changes:
- Refactor greyScale() to output float [0,1] instead of int32 [0,65535]
- Update all function signatures: int32_t* → float* for image data
- Adapt buffers: g_grayScale_live, processed_grayScale, g_frozen_grayscale_buffer
- Update process_image_preprocessing() to accept float input
- Update calculate_contrast() and synth_precompute_wave_data() signatures
- Adapt worker thread imageData type to float*

This fix ensures consistent normalized values throughout the audio pipeline
and resolves saturation issues on Raspberry Pi with high exponent values.

+++ src/synthesis/additive/synth_additive_algorithms.h
@@ -26,8 +26,8 @@
  * @param end_note Ending note index (exclusive)
  * @retval None
  */
-void process_image_preprocessing(int32_t *imageData, int32_t *imageBuffer_q31, 
-                                int start_note, int end_note);
+void process_image_preprocessing(float *imageData, int32_t *imageBuffer_q31, 
+                                 int start_note, int end_note);
 
 /**
  * @brief Apply GAP_LIMITER volume ramp for a single note
